requirements_websocket.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
opencv-python==4.11.0.86
numpy==1.26.4
python-engineio==4.7.1
